Add repeated_nodes cleanup, timing stats
The repeated_nodes list is now cleaned up every 10 simulation timesteps.
This reduces the amount of data stored in the RAM, and should also
slightly increase the performance of the _random_subset-function.
Additionally, timing information is collected and printed to stdout
to check how the timing scales with the number of users.